+++ test/validation-errors.js
@@ -1,11 +1,18 @@
-const { expect } = require( 'chai' );
+const { expectValidateFixturesResult } = require( './lib' );
 
-const { runValidationWithFixtures } = require( './lib' );
+describe( 'validation success', () => {
+	it( 'should succeed with valid data', () => {
+		expectValidateFixturesResult( 'valid', {
+			errorCount: 0,
+			output: [],
+		} );
+	} );
+} );
 
-describe( 'validation script errors', () => {
+describe( 'validation errors', () => {
 	it( 'should catch invalid table rows', () => {
-		expect( runValidationWithFixtures( 'bad-table-rows' ) ).to.eql( {
-			exitCode: 2,
+		expectValidateFixturesResult( 'bad-table-rows', {
+			errorCount: 2,
 			output: [
 				'README.md: Expected 3 table cells but found 2: <td><a href="/company-profiles/10up.md">10up</a></td><td><a href="https://10up.com/">https://10up.com/</a></td>',
 				'README.md: Expected 3 table cells but found 4: <td><a href="/company-profiles/18f.md">18F</a></td><td><a href="https://18f.gsa.gov/">https://18f.gsa.gov/</a></td><td>USA</td><td>something else</td>',
@@ -14,8 +21,8 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch missing company names', () => {
-		expect( runValidationWithFixtures( 'missing-company-names' ) ).to.eql( {
-			exitCode: 11,
+		expectValidateFixturesResult( 'missing-company-names', {
+			errorCount: 11,
 			output: [
 				'README.md: Company "⚠⚠⚠" has no linked Markdown profile (".md")',
 				'README.md: Missing company name: <td></td><td><a href="https://andyet.com">https://andyet.com</a></td><td>Worldwide</td>',
@@ -33,8 +40,8 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch unsorted company names', () => {
-		expect( runValidationWithFixtures( 'unsorted' ) ).to.eql( {
-			exitCode: 2,
+		expectValidateFixturesResult( 'unsorted', {
+			errorCount: 2,
 			output: [
 				'README.md: Company is listed out of order: "17hats" (should be before "18F")',
 				'README.md: Company is listed out of order: "&yet" (should be before "17hats")',
@@ -43,8 +50,8 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch invalid profile links and missing profiles', () => {
-		expect( runValidationWithFixtures( 'bad-profile-links' ) ).to.eql( {
-			exitCode: 4,
+		expectValidateFixturesResult( 'bad-profile-links', {
+			errorCount: 4,
 			output: [
 				'README.md: Invalid link to company "&yet": "company-profiles/and-yet.md"',
 				'README.md: Broken link to company "17hats": "/company-profiles/17hats-nonexistent.md"',
@@ -55,8 +62,8 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch invalid titles in company profiles', () => {
-		expect( runValidationWithFixtures( 'bad-profile-titles' ) ).to.eql( {
-			exitCode: 7,
+		expectValidateFixturesResult( 'bad-profile-titles', {
+			errorCount: 7,
 			output: [
 				'10up.md: Expected 1 first-level heading but found 0',
 				'10up.md: The main title is wrapped inside of another element.',
@@ -70,8 +77,8 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch orphaned company profiles', () => {
-		expect( runValidationWithFixtures( 'orphaned-profiles' ) ).to.eql( {
-			exitCode: 1,
+		expectValidateFixturesResult( 'orphaned-profiles', {
+			errorCount: 1,
 			output: [
 				'18f.md: No link to company profile from readme',
 			],
@@ -79,8 +86,8 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch invalid section headings', () => {
-		expect( runValidationWithFixtures( 'bad-profile-headings' ) ).to.eql( {
-			exitCode: 10,
+		expectValidateFixturesResult( 'bad-profile-headings', {
+			errorCount: 10,
 			output: [
 				'10up.md: Required section "Company blurb" not found.',
 				'17hats.md: Invalid section: "A thing I made up".  Expected one of: ["Company blurb","Company size","Remote status","Region","Company technologies","Office locations","How to apply"]',
@@ -97,8 +104,8 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch text outside of links in readme', () => {
-		expect( runValidationWithFixtures( 'name-outside-link' ) ).to.eql( {
-			exitCode: 3,
+		expectValidateFixturesResult( 'name-outside-link', {
+			errorCount: 3,
 			output: [
 				'README.md: Extra text in company name: "10up", "10up agency"',
 				'README.md: Extra text in company name: "Aerolab", "Aerolab  more text"',
@@ -108,14 +115,16 @@ describe( 'validation script errors', () => {
 	} );
 
 	it( 'should catch mismatched "incomplete profile" indicators', () => {
-		expect( runValidationWithFixtures( 'mismatched-incomplete-indicators' ) ).to.eql( {
-			exitCode: 7,
+		expectValidateFixturesResult( 'mismatched-incomplete-indicators', {
+			errorCount: 7,
 			output: [
 				'10up.md: Profile is marked as complete, but it only contains a "Company blurb" heading.',
 				'17hats.md: Profile looks complete, but the "Company blurb" contains a warning emoji.',
 				'18f.md: Profile looks incomplete, but the main readme does not contain a warning emoji.',
-				'45royale.md: Profile is marked as incomplete, but it contains multiple sections.',
-				'45royale.md: Please remove the warning emoji from the "Company blurb" section and the main readme.',
+				(
+					'45royale.md: Profile is marked as incomplete, but it contains multiple sections.'
+					+ '\nPlease remove the warning emoji from the "Company blurb" section and the main readme.'
+				),
 				'aerolab.md: Profile looks incomplete, but the "Company blurb" does not contain a warning emoji.',
 				'and-yet.md: Profile looks complete, but the main readme contains a warning emoji.',
 				'angularclass.md: Profile looks incomplete, but the "Company blurb" does not contain a warning emoji.',
